add class balance same as in datapatch layer to  softmaxloss layer
move lable_select_param to layer_param level

+++ src/caffe/util/im2col.cpp
@@ -47,19 +47,19 @@ inline void im2col_nd_core_cpu(const Dtype* data_input, const bool im2col,
     Dtype* data_output) {
   if (!im2col) {
     //int im_size = im_shape[0];
-    size_t im_size = im_shape[0];
+    long im_size = im_shape[0];
     for (int i = 0; i < num_spatial_axes; ++i) {
-      im_size *= (size_t)im_shape[1 + i];
+      im_size *= (long)im_shape[1 + i];
     }
     caffe_set(im_size, Dtype(0), data_output);
   }
-  int kernel_size = 1;
+  long kernel_size = 1;
   for (int i = 0; i < num_spatial_axes; ++i) {
     kernel_size *= kernel_shape[i];
   }
   const int channels_col = col_shape[0];
-  vector<int> d_offset(num_spatial_axes, 0);
-  vector<int> d_iter(num_spatial_axes, 0);
+  vector<long> d_offset(num_spatial_axes, 0);
+  vector<long> d_iter(num_spatial_axes, 0);
   for (int c_col = 0; c_col < channels_col; ++c_col) {
     // Loop over spatial axes in reverse order to compute a per-axis offset.
     int offset = c_col;
@@ -74,19 +74,19 @@ inline void im2col_nd_core_cpu(const Dtype* data_input, const bool im2col,
       // image and column, and whether the index lies in the padding.
     //  int index_col = c_col;
     //  int index_im = c_col / kernel_size;
-      size_t index_col = c_col;
-      size_t index_im = c_col / kernel_size;
+      long index_col = c_col;
+      long index_im = c_col / kernel_size;
       bool is_padding = false;
       for (int d_i = 0; d_i < num_spatial_axes; ++d_i) {
         //const int d = d_iter[d_i];
         //const int d_im = d * stride[d_i] - pad[d_i] + d_offset[d_i];
-        const size_t d = d_iter[d_i];
-        const size_t d_im = d * stride[d_i] - pad[d_i] + d_offset[d_i];
+        const long d = d_iter[d_i];
+        const long d_im = d * stride[d_i] - pad[d_i] + d_offset[d_i];
         is_padding |= d_im < 0 || d_im >= im_shape[d_i + 1];
-        index_col *= (size_t)col_shape[d_i + 1];
-        index_col += (size_t)d;
-        index_im *= (size_t)im_shape[d_i + 1];
-        index_im += (size_t)d_im;
+        index_col *= (long)col_shape[d_i + 1];
+        index_col += (long)d;
+        index_im *= (long)im_shape[d_i + 1];
+        index_im += (long)d_im;
       }
       if (im2col) {
         if (is_padding) {
